Dashboard started charts done

+++ serverapi/Models/SystemData.cs
@@ -4,8 +4,8 @@ namespace serverapi.Models
 {
     public class SystemData
     {
-        public double CpuPercent { get; }
-        public double MemoryMb { get; }
+        public float CpuPercent { get; }
+        public float MemoryMb { get; }
 
         public SystemData(ContainerStatsResponse stats)
         {
@@ -17,14 +17,14 @@ namespace serverapi.Models
 
             CpuPercent = systemDelta == 0
                 ? 0
-                : Math.Round((double)cpuDelta / systemDelta * numCpus * 100.0, 2);
+                : (float)Math.Round((float)cpuDelta / systemDelta * numCpus * 100.0, 2);
 
             var cache = stats.MemoryStats.Stats.ContainsKey("cache")
                 ? stats.MemoryStats.Stats["cache"]
                 : 0;
             var usedMemory = stats.MemoryStats.Usage - cache;
 
-            MemoryMb = Math.Round(usedMemory / 1024.0 / 1024.0, 2);
+            MemoryMb = (float)Math.Round(usedMemory / 1024.0 / 1024.0, 2);
         }
     }
 }
